added invoice configurations
