+++ src/Form/Model/SystemConfiguration.php
@@ -13,6 +13,7 @@ class SystemConfiguration
 {
     public const SECTION_ROUNDING = 'rounding';
     public const SECTION_TIMESHEET = 'timesheet';
+    public const SECTION_FORM_INVOICE = 'invoice';
     public const SECTION_FORM_CUSTOMER = 'form_customer';
     public const SECTION_FORM_USER = 'form_user';
     public const SECTION_THEME = 'theme';
